requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
mysql-connector==2.2.9
mysqlclient==1.4.3
numpy==1.17.0
opencv-python==4.1.0.25
pkg-resources==0.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
tqdm==4.33.0
